Fix location id building level 3 - SL -
Add country location code awareness

Adapted SL location with this requirements.
In Sierra Leona we have the following problem:
level 2 can be with padding '01' or without padding '1'
level 3 has to be built from the level 3 value and the level 2 value
i.e. if level 2 is '01' and level 3 is '34' the resulting id should be
1-34.